Store paramRef to recalculate edges
Previously, if the paramRef name, namespace, or selector were changed,
the edges were not recalculated because no properties on the node were
changing. Now the reconciler specially checks the metadata on VAPBs for
those kinds of changes.

A reconciler test was updated to handle update events, to observe this
behavior.

Refs:
 - https://issues.redhat.com/browse/ACM-16246

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>